Edge Server/requirements.txt
absl-py==1.3.0
asttokens==2.1.0
astunparse==1.6.3
backcall==0.2.0
cachetools==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.4
contourpy==1.0.6
cycler==0.11.0
dask==2022.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.4
dm-tree==0.1.8
dnspython==2.3.0
entrypoints==0.4
etils==0.9.0
executing==1.2.0
filelock==3.7.0
flatbuffers==23.5.26
fonttools==4.38.0
fsspec==2022.11.0
future==0.18.2
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.57.0
grpcio==1.50.0
h5py==3.7.0
httpie==3.0.2
idna==3.3
importlib-metadata==5.0.0
importlib-resources==5.10.1
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.1
joblib==1.2.0
jupyter_client==7.4.6
jupyter_core==5.0.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
locket==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
multidict==6.0.2
nest-asyncio==1.5.6
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
paho-mqtt==1.6.1
panda==0.3.1
pandas==1.5.1
parso==0.8.3
partd==1.3.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.2
promise==2.3
prompt-toolkit==3.0.32
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.2
Pygments==2.11.2
pymongo==4.3.3
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.6
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
six==1.15.0
sklearn==0.0.post1
stack-data==0.6.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-datasets==4.8.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-metadata==1.12.0
termcolor==1.1.0
tflite==2.10.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch==1.5.1+cu101
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing-extensions==3.7.4.3
urllib3==1.26.8
virtualenv==20.14.1
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.12.1
zipp==3.10.0
